Compiled and formatted JMA and WOD data
This is the compiled and formatted JMA and WOD data in the paper entitled "Temporal and spatial variations of cross-shelf nutrient exchange in the East China Sea, as estimated by satellite altimetry and in situ measurements". They have been prepared as mat file, which can be read using MATLAB. For JMA data (named as jma),  columns 1 to 201 denote the nitrate data from 0 to 200 m, and 202 to 205 denote longitude, latitude, month and year. For WOD data (named as wod), the first dimension (4) is time of month (1, 4, 7, 10), and the second dimension (201) is depth from 0 to 200 m, and the third dimension (200) is section points at the 200 m isobath. The location of section points can refer to 'lo' and 'la' mat file.